refactor(middleware): streamline redirect handling and introduce API for fetching redirects
- Removed the Sanity client initialization from the middleware, replacing it with a fetch call to a new API endpoint for redirects.
- Simplified the redirect map structure to use a Record instead of a Map for easier access.
- Added a new API route to handle fetching redirect entries from Sanity, improving separation of concerns and maintainability.
- Updated error handling in the middleware to ensure graceful failure when fetching redirects fails.
